+++ app/src/main/java/no/agens/darjeeling/AnotherActivity.kt
@@ -1,0 +1,13 @@
+package no.agens.darjeeling
+
+import android.app.Activity
+import android.os.Bundle
+
+class AnotherActivity : Activity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_another)
+    }
+
+}